src/main/java/com/webtoon/comment/application/CommentService.java
package com.webtoon.comment.application;

import com.webtoon.comment.domain.Comment;
import com.webtoon.comment.dto.CommentSave;
import com.webtoon.comment.repository.CommentRepository;
import com.webtoon.content.domain.Content;
import com.webtoon.content.repository.ContentRepository;
import com.webtoon.member.domain.Member;
import com.webtoon.member.repository.MemberRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@RequiredArgsConstructor
@Transactional(readOnly = true)
@Service
public class CommentService {

    private final CommentRepository commentRepository;
    private final MemberRepository memberRepository;
    private final ContentRepository contentRepository;

    @Transactional
    public Comment save(CommentSave commentSave, Long memberSessionId, Long contentId) {
        Member member = memberRepository.getById(memberSessionId);
        Content content = contentRepository.getById(contentId);
        Comment comment = commentSave.toEntity(member, content);
        return commentRepository.save(comment);
    }
}
